Now resolving match expressions before compilation
